+++ script/build_and_run.sh
@@ -66,9 +66,9 @@ cat >"$INFO_PLIST" <<PLIST
   <key>CFBundleName</key>
   <string>$APP_NAME</string>
   <key>CFBundleShortVersionString</key>
-  <string>1.8.2</string>
+  <string>1.8.4</string>
   <key>CFBundleVersion</key>
-  <string>1.8.2</string>
+  <string>1.8.4</string>
   <key>CFBundlePackageType</key>
   <string>APPL</string>
   <key>LSMinimumSystemVersion</key>